Set libpcap to compile against SDK 29.
This is needed to make NetworkStackTests JNI code compile against
the NDK so it can run on both Q and future releases.

The lowest NDK version that libpcap builds for is 24 (because it
needs getifaddrs). However, using 29 decreases future maintenance
burden, because if we use 24 and upstream libpcap later on starts
needing a symbol that was only introduced in, say, 28, we'd have
to update all the sdk_versions of the users of libpcap, or fork.

libpcap is external code, so it is not likely that it will ever
need to depend on Android platform internals.

Bug: 148753875
Test: m libpcap
Test: atest NetworkStackTests
Change-Id: I708c4d0d8096b37f960b66e11ec00a1b8d1fb83d